third_party/blink/web_tests/fast/scrolling/events/scrollend-event-fired-to-document.html

<!DOCTYPE HTML>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/gesture-util.js"></script>
<style>
#targetDiv {
  width: 200px;
  height: 200px;
  overflow: scroll;
}

#innerDiv {
  width: 400px;
  height: 400px;
}
</style>

<body style="margin:0">
<div id="targetDiv">
  <div id="innerDiv">
  </div>
</div>
</body>

<script>
if (window.internals)
  internals.runtimeFlags.overscrollCustomizationEnabled = true;

var target_div = document.getElementById('targetDiv');
var rect = target_div.getBoundingClientRect();
var x = rect.left + rect.width / 2;
var y = rect.top + rect.height / 2;
function scrollLeft() {
  return smoothScroll(1000, x, y, GestureSourceType.TOUCH_INPUT, "left",
      SPEED_INSTANT);
}

function scrollUp() {
  return smoothScroll(1000, x, y, GestureSourceType.TOUCH_INPUT, "up",
      SPEED_INSTANT);
}

var horizontal_scrollend_arrived = false;
var vertical_scrollend_arrived = false;
function onHorizontalScrollEnd(event) {
  assert_false(event.cancelable);
  assert_false(event.bubbles);
  horizontal_scrollend_arrived = true;
}
function onVerticalScrollEnd(event) {
  assert_false(event.cancelable);
  assert_false(event.bubbles);
  vertical_scrollend_arrived = true;
}
document.addEventListener("scrollend", onHorizontalScrollEnd);
document.addEventListener("scrollend", onVerticalScrollEnd);

promise_test (async (t) => {
  // Make sure that no scrollend event is sent to target_div.
  target_div.addEventListener("scrollend",
      t.unreached_func("target_div got unexpected scrollend event."));

  await waitForCompositorCommit();

  // Scroll left on target div and wait for the doc to get scrollend event.
  await scrollLeft();
  await waitFor(() => { return horizontal_scrollend_arrived; },
      'Document did not receive scrollend event after scroll left on target.');
  assert_equals(target_div.scrollLeft, 0);

  // Scroll up on target div and wait for the doc to get scrollend event.
  await scrollUp();
  await waitFor(() => { return horizontal_scrollend_arrived; },
      'Document did not receive scrollend event after scroll up on target.');
  assert_equals(target_div.scrollTop, 0);
}, 'Tests that the document gets scrollend event when no element scrolls.');
</script>
